pyproject.toml
[tool.poetry]
name = "gha-stats"
version = "0.1.0"
description = ""
authors = ["Paul Gessinger <hello@paulgessinger.com>"]
packages = [
    { include = "gha_stats", from = "src" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
aiohttp = "^3.7.4"
gidgethub = "^5.0.1"
typer = "^0.4.1"
peewee = "^3.14.10"
pysqlite3 = "^0.4.7"
cachetools = "^5.1.0"
more-itertools = "^8.13.0"
aioitertools = "^0.10.0"
python-dateutil = "^2.8.2"

[tool.poetry.scripts]
gha-stats = 'gha_stats.cli:cli'

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
